requirements.txt
telethon>=1.29.0
PyYAML>=6.0
requests>=2.31.0
aiohttp>=3.8.5
pyinstaller>=6.4.0
